backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
Django==5.2.6
django-cors-headers==4.8.0
django-filter==25.1
django-otp==1.6.3
djangorestframework==3.16.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.77
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.27
msgpack==1.1.1
openai==2.1.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==1.7.1
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
stripe==12.5.1
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.24.0
python-dateutil
gunicorn==21.2.0
psycopg2-binary>=2.9.9
djangorestframework-api-key
